fix : changes done as suggested in comments
