+++ erpnext/selling/report/unmet_customer_sales_targets/unmet_customer_sales_targets.js
@@ -2,6 +2,8 @@
 // For license information, please see license.txt
 /* eslint-disable */
 
-frappe.query_reports["Customer sales target not met"] = {
-	"filters": []
+frappe.query_reports["Unmet Customer Sales Targets"] = {
+	"filters": [
+
+	]
 };
