+++ erpnext/selling/report/unmet_customer_sales_targets/unmet_customer_sales_targets.py
@@ -1,0 +1,49 @@
+# Copyright (c) 2013, Frappe Technologies Pvt. Ltd. and contributors
+# For license information, please see license.txt
+
+from __future__ import unicode_literals
+from frappe import _
+import frappe
+
+def execute(filters=None):
+	columns=get_columns(filters)
+
+	data = get_data(filters)
+
+	return columns, data
+
+def get_columns(filters):
+	return [
+		{
+			"fieldname": "customer_name",
+			"label": _("Customer"),
+			"fieldtype": "Link",
+			"options": "Customer",
+			"width": 200
+		},
+		{
+			"fieldname": "monthly_sales_target",
+			"label": _("Monthly Sales Target"),
+			"fieldtype": "Currency",
+			"width": 150
+		},
+		{
+			"fieldname": "total_monthly_sales",
+			"label": _("Total Monthly Sales"),
+			"fieldtype": "Currency",
+			"width": 150
+		}
+	]
+
+def get_data(filters= None):
+	return frappe.db.sql("""
+		SELECT
+		customer_name,
+			monthly_sales_target,
+			total_monthly_sales
+        FROM
+			tabCustomer
+    	WHERE
+			monthly_sales_target > total_monthly_sales
+		""")
+
